Update routing to always set token from the tokenValidation query parameter
