+++ src/routes.js
@@ -7,7 +7,7 @@ const routesArray = [
   { path: '/compare/:ids', key: 'compareID', componentName: 'Compare', pageTitle: 'Compare Positions' },
   { path: '/compare', key: 'compareNoID', componentName: 'Compare', pageTitle: 'Compare Positions' },
   { path: '/about', exact: true, componentName: 'About', pageTitle: 'About' },
-  { path: '/tokenValidation', componentName: 'Login', pageTitle: 'Token Validation' },
+  { path: '/tokenValidation', componentName: 'TokenValidation', pageTitle: 'Token Validation' },
   { path: '/loginRedirect', componentName: 'LoginRedirect', pageTitle: 'Login Redirect' },
 ];
